+++ src/main/java/nextstep/qna/exception/UnAuthorizedException.java
@@ -1,4 +1,4 @@
-package nextstep.qna;
+package nextstep.qna.exception;
 
 public class UnAuthorizedException extends RuntimeException {
     private static final long serialVersionUID = 1L;
