+++ src/main/java/nextstep/qna/service/QnAService.java
@@ -1,7 +1,7 @@
 package nextstep.qna.service;
 
-import nextstep.qna.CannotDeleteException;
-import nextstep.qna.NotFoundException;
+import nextstep.qna.exception.CannotDeleteException;
+import nextstep.qna.exception.NotFoundException;
 import nextstep.qna.domain.AnswerRepository;
 import nextstep.qna.domain.DeleteHistories;
 import nextstep.qna.domain.Question;
